+++ src/main/c/jsonpath/subscript.c
@@ -1,7 +1,7 @@
 #include "jsonpath.h"
 #include "conditions.h"
 
-size_t subscript_predicate_index(const Predicate *value)
+int64_t subscript_predicate_index(const Predicate *value)
 {
     PRECOND_NONNULL_ELSE_ZERO(value);
     PRECOND_ELSE_ZERO(SUBSCRIPT == value->kind);
